requirements.txt
requests
beautifulsoup4
flask
google-generativeai
